Debug Exception Routines
Make debug exception routines weak so that a module can
override them. These exceptions are indeed supposed to be
used to debug and debugging requires saving much more
context to be able to send it to the client and making
possible to modify this context from the client interface.

For now, the new handler functions are the same as core's
but will be soonly changed to save the entire CPU context.