Adding minimal code and data mapping in tasks address spaces.
Fixing tests that make the assumption of the kernel code and data being fully
mapped in non-kernel tasks.
For these tests, the shared variables are now placed at a fixed virtual address
within the tasks address spaces. An improvement could be to write to the
low-level context structure to pass the address to the threads main function.